Move Connection and ConnectionSettings classes from api to client module.
