+++ components/client/src/main/java/com/hotels/styx/client/Connection.java
@@ -13,10 +13,11 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.client;
+package com.hotels.styx.client;
 
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponse;
+import com.hotels.styx.api.client.Origin;
 import rx.Observable;
 
 import java.io.Closeable;
