+++ components/client/src/main/java/com/hotels/styx/client/ConnectionSettings.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.client;
+package com.hotels.styx.client;
 
 import com.hotels.styx.api.service.TlsSettings;
