+++ components/client/src/main/java/com/hotels/styx/client/Operation.java
@@ -15,7 +15,6 @@
  */
 package com.hotels.styx.client;
 
-import com.hotels.styx.api.client.Connection;
 import rx.Observable;
 
 /**
