+++ components/client/src/main/java/com/hotels/styx/client/SimpleHttpClient.java
@@ -21,8 +21,6 @@ import com.hotels.styx.api.FullHttpClient;
 import com.hotels.styx.api.FullHttpRequest;
 import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.Url;
-import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionSettings;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.service.TlsSettings;
 import com.hotels.styx.client.netty.connectionpool.HttpRequestOperation;
@@ -48,7 +46,7 @@ public final class SimpleHttpClient implements FullHttpClient {
     private static final int DEFAULT_HTTP_PORT = 80;
 
     private final Optional<String> userAgent;
-    private final com.hotels.styx.api.client.ConnectionSettings connectionSettings;
+    private final ConnectionSettings connectionSettings;
     private final int maxResponseSize;
     private final Connection.Factory connectionFactory;
 
@@ -104,13 +102,13 @@ public final class SimpleHttpClient implements FullHttpClient {
         private Connection.Factory connectionFactory;
         private TlsSettings tlsSettings;
         private String userAgent;
-        private com.hotels.styx.api.client.ConnectionSettings connectionSettings = new ConnectionSettings(1000);
+        private ConnectionSettings connectionSettings = new ConnectionSettings(1000);
         private int responseTimeout = 60000;
         private int maxResponseSize = 1024 * 100;
         private int maxHeaderSize = 8192;
         private String threadName = "simple-netty-http-client";
 
-        public Builder connectionSettings(com.hotels.styx.api.client.ConnectionSettings connectionSettings) {
+        public Builder connectionSettings(ConnectionSettings connectionSettings) {
             this.connectionSettings = connectionSettings;
             return this;
         }
