+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ConnectionPool.java
@@ -15,7 +15,7 @@
  */
 package com.hotels.styx.client.connectionpool;
 
-import com.hotels.styx.api.client.Connection;
+import com.hotels.styx.client.Connection;
 import com.hotels.styx.api.client.Origin;
 import rx.Observable;
