+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ConnectionPoolFactory.java
@@ -15,7 +15,7 @@
  */
 package com.hotels.styx.client.connectionpool;
 
-import com.hotels.styx.api.client.Connection;
+import com.hotels.styx.client.Connection;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.metrics.MetricRegistry;
 import com.hotels.styx.api.service.ConnectionPoolSettings;
