+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ExpiringConnection.java
@@ -19,7 +19,7 @@ import com.google.common.base.Stopwatch;
 import com.google.common.base.Ticker;
 import com.hotels.styx.api.HttpRequest;
 import com.hotels.styx.api.HttpResponse;
-import com.hotels.styx.api.client.Connection;
+import com.hotels.styx.client.Connection;
 import com.hotels.styx.api.client.Origin;
 import rx.Observable;
