+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ExpiringConnectionFactory.java
@@ -16,8 +16,8 @@
 package com.hotels.styx.client.connectionpool;
 
 import com.google.common.base.Ticker;
-import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionSettings;
+import com.hotels.styx.client.Connection;
+import com.hotels.styx.client.ConnectionSettings;
 import com.hotels.styx.api.client.Origin;
 import rx.Observable;
