+++ components/client/src/main/java/com/hotels/styx/client/netty/connectionpool/NettyConnectionFactory.java
@@ -16,8 +16,8 @@
 package com.hotels.styx.client.netty.connectionpool;
 
 import com.google.common.net.HostAndPort;
-import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionSettings;
+import com.hotels.styx.client.Connection;
+import com.hotels.styx.client.ConnectionSettings;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.netty.ClientEventLoopFactory;
 import com.hotels.styx.api.exceptions.OriginUnreachableException;
