+++ components/proxy/src/test/java/com/hotels/styx/admin/tasks/StubConnectionPool.java
@@ -15,7 +15,7 @@
  */
 package com.hotels.styx.admin.tasks;
 
-import com.hotels.styx.api.client.Connection;
+import com.hotels.styx.client.Connection;
 import com.hotels.styx.client.connectionpool.ConnectionPool;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.service.ConnectionPoolSettings;
